Agregué validaciones a la dao en la que trabajo y excepciones a todo el proyecto

+++ ObjetoNegocio/src/main/java/org/itson/bdavanzadas/ObjetoNegocio/ProductoBO.java
@@ -8,6 +8,7 @@ import org.itson.bdavanzadas.persistencia.conexion.Conexion;
 import org.itson.bdavanzadas.persistencia.conexion.IConexion;
 import org.itson.bdavanzadas.persistencia.daos.IProductoDAO;
 import org.itson.bdavanzadas.persistencia.daos.ProductoDAO;
+import org.itson.bdavanzadas.persistencia.exception.PersistenciaException;
 
 /**
  *
@@ -22,17 +23,21 @@ public class ProductoBO {
         productoDAO = new ProductoDAO(conexion);
     }
 
-    public void agregarNuevoProducto(String nombre, String descripcion, double precio, int cantidad, String categoria) {
+    public void agregarNuevoProducto(String nombre, String descripcion, double precio, int cantidad, String categoria) throws PersistenciaException {
         if (!productoDAO.productoExiste(nombre)) {
             productoDAO.agregarNuevoProducto(nombre, descripcion, precio, cantidad, categoria);
         }
     }
 
-    public void actualizarProducto(String nombrePrevio, String nuevoNombre, String nuevaDescripcion, double nuevoPrecio, int nuevaCantidad, String nuevaCategoria) {
-        productoDAO.actualizarProducto(nombrePrevio, nuevoNombre, nuevaDescripcion, nuevoPrecio, nuevaCantidad, nuevaCategoria);
+    public void actualizarProducto(String nombrePrevio, String nuevoNombre, String nuevaDescripcion, double nuevoPrecio, int nuevaCantidad, String nuevaCategoria) throws PersistenciaException {
+        if (productoDAO.productoExiste(nombrePrevio)) {
+            productoDAO.actualizarProducto(nombrePrevio, nuevoNombre, nuevaDescripcion, nuevoPrecio, nuevaCantidad, nuevaCategoria);
+        }
     }
 
-    public void eliminarProducto(String nombre) {
-        productoDAO.eliminarProducto(nombre);
+    public void eliminarProducto(String nombre) throws PersistenciaException {
+        if (productoDAO.productoExiste(nombre)) {
+            productoDAO.eliminarProducto(nombre);
+        }
     }
 }
